Fix graphic queue bugs (#31)
* Refactored and fixed states

* Fixed stunned clients bug

* Refactored and fixed states

* Fixed stunned clients bug

* Fixes
