+++ src/main/java/ua/pz33/clients/Client.java
@@ -31,7 +31,7 @@ public class Client implements ClockObserver {
     }
 
     public void buyTickets(CashRegister cashRegister) {
-        var message = String.format("Client %d bought %d tickets in cashregister %d", id, countOfTickets, cashRegister.getId());
+        var message = String.format("Client %d bought %d tickets in %d cashregister.", id, countOfTickets, cashRegister.getId());
         countOfTickets = 0;
         LogMediator.getInstance().logMessage(message);
     }
@@ -55,7 +55,6 @@ public class Client implements ClockObserver {
                 ? StationController.getInstance().getBackupCashRegisterSprite(c.getId())
                 : StationController.getInstance().getCashRegisterSprite(c.getId())));
 
-        //todo choose the closest
         var closestCashRegisterSprite = openedCashRegisterSprites.stream().min(closestCashRegisterComparator).get();
         var bestCashRegister = isBackup
                 ? StationController.getInstance().getBackupCashRegister(closestCashRegisterSprite.getId())
