+++ src/main/java/ua/pz33/clients/statemachice/ExitState.java
@@ -4,6 +4,7 @@ import ua.pz33.StationController;
 import ua.pz33.clients.Client;
 
 public class ExitState extends State{
+    private boolean executedAnim = false;
     public ExitState(Client client) {
         super(client);
     }
@@ -12,6 +13,11 @@ public class ExitState extends State{
     public void perform() {
         var clientSprite = StationController.getInstance().getClientSprite(client.getId());
         var goalPoint = client.getGoalPoint();
+
+        if(!executedAnim){
+            clientSprite.moveTo(goalPoint.x, goalPoint.y);
+            executedAnim = true;
+        }
 
         //we arrived exit
         if (clientSprite.getX() == goalPoint.x && clientSprite.getY() == goalPoint.y) {
